infrastructure-db/discovery/requirements.txt
# Python dependencies for infrastructure discovery scripts
# Install with: pip install -r requirements.txt

# SSH and remote execution
paramiko>=3.0.0

# API clients
requests>=2.31.0
proxmoxer>=2.0.0

# Note: sqlite3 is built-in to Python, no need to install

# CLI utilities
rich>=13.0.0  # Beautiful terminal output
python-dotenv>=1.0.0  # Environment variable management
click>=8.1.0  # CLI framework

# JSON/YAML parsing
pyyaml>=6.0

# Date/time utilities
python-dateutil>=2.8.0
